+++ modules/Users/User.php
@@ -321,7 +321,7 @@ if ( $_REQUEST['modfunc'] === 'update'
 							continue;
 						}
 
-						$value = encrypt_password( str_replace( "''", "'", $value ) );
+						$value = encrypt_password( $_POST['staff']['PASSWORD'] );
 					}
 
 					$sql .= DBEscapeIdentifier( $column ) . "='" . $value . "',";
@@ -399,7 +399,7 @@ if ( $_REQUEST['modfunc'] === 'update'
 					}
 					elseif ( $column == 'PASSWORD' )
 					{
-						$value = encrypt_password( str_replace( "''", "'", $value ) );
+						$value = encrypt_password( $_POST['staff']['PASSWORD'] );
 					}
 
 					if ( ! empty( $value ) || $value == '0' )